test: fix installation failure caused by puppeteer
