+++ packages/@vue/cli-plugin-babel/__tests__/babelMigrator.spec.js
@@ -5,6 +5,9 @@ jest.setTimeout(300000)
 beforeEach(() => {
   process.env.VUE_CLI_TEST_DO_INSTALL_PLUGIN = true
 })
+afterAll(() => {
+  delete process.env.VUE_CLI_TEST_DO_INSTALL_PLUGIN
+})
 
 test('upgrade: plugin-babel v3.5', async () => {
   const project = await create('plugin-babel-legacy', {
