+++ packages/@vue/cli/lib/util/ProjectPackageManager.js
@@ -301,7 +301,9 @@ class PackageManager {
   async install () {
     if (process.env.VUE_CLI_TEST) {
       try {
+        process.env.PUPPETEER_SKIP_CHROMIUM_DOWNLOAD = true
         await this.runCommand('install', ['--offline', '--silent', '--no-progress'])
+        delete process.env.PUPPETEER_SKIP_CHROMIUM_DOWNLOAD
       } catch (e) {
         await this.runCommand('install', ['--silent', '--no-progress'])
       }
